Change needed to make chrome tests build again
There was a commit (75c784ae54225401740d014f8d2978b266926939) which added
UnitTestsQWeb to the ChromeTests project.
That ultimately references the QServer project.
This has a after build step that will try and copy some files in a particular location based
on configuration and platform.

These two parameters need to be set correctly in order for things to work.

When building through VS it's all fine.
When building a sln through msbuild it's also fine.
However, when building a csproj through msbuild it's not fine.

For that case, we just need to call msbuild with the right parameters to make it happy.